Splitte søknad/index i 3 filer, da den har blitt svært stor og lite lettlest
